# backend-api/doc/openapiSpec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shopping Website API",
    "description": "REST API cho hệ thống thương mại điện tử – Node.js / Express / PostgreSQL",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false,
            "description": "Luôn là false khi có lỗi"
          },
          "message": {
            "type": "string",
            "description": "Nhãn bao quát: 'Bad request' (4xx) hoặc 'Internal server error' (5xx)",
            "example": "Bad request"
          },
          "error": {
            "type": "string",
            "description": "Chi tiết lỗi hiển thị cho client. 4xx: message thật từ service. 5xx: 'Hệ thống gặp sự cố, vui lòng thử lại sau.'",
            "example": "Chỉ còn 2 sản phẩm trong kho"
          },
          "stack": {
            "type": "string",
            "description": "Stack trace — chỉ có trong môi trường development",
            "nullable": true
          }
        },
        "required": ["success", "message", "error"]
      },
...
      "Blog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "is_deleted": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "thumbnail_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          }
        }
      },
      "Contact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "message": {
            "type": "string"
          },
          "user_id": {
            "type": "integer",
            "nullable": true
          },
          "status": {
            "type": "string",
            "example": "pending"
          },
          "reply": {
            "type": "string",
            "nullable": true
          },
          "replied_by": {
            "type": "integer",
            "nullable": true
          },
          "replied_by_name": {
            "type": "string",
            "nullable": true,
            "description": "Tên staff/admin đã phản hồi (được lấy từ bảng users)"
          },
          "replied_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
      "IdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Chưa xác thực — thiếu hoặc token không hợp lệ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Bad request",
              "error": "No token provided"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Không có quyền truy cập",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Bad request",
              "error": "Cart item not found"
            }
          }
        }
      },
      "NotFound": {
        "description": "Không tìm thấy tài nguyên",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Bad request",
              "error": "Order not found"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Dữ liệu đầu vào không hợp lệ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Bad request",
              "error": "product_id: Required"
            }
          }
        }
      },
      "InternalError": {
        "description": "Lỗi hệ thống",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Internal server error",
              "error": "Hệ thống gặp sự cố, vui lòng thử lại sau."
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/blogs": {
      "get": {
        "tags": ["Blogs"],
        "summary": "Lấy tất cả bài viết blog (Public)",
        "responses": {
          "200": {
            "description": "Danh sách blog",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Blog"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "operationId": "listBlogs"
      },
      "post": {
        "tags": ["Blogs"],
        "summary": "Tạo bài viết blog (Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "thumbnail_url": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL ảnh đại diện bài viết"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bài viết đã tạo",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Blog"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "operationId": "createBlog"
      }
    },
...
    "/api/blogs/{id}": {
      "put": {
        "tags": ["Blogs"],
        "summary": "Cập nhật bài viết blog (Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "thumbnail_url": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL ảnh đại diện bài viết"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã cập nhật",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Blog"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "operationId": "updateBlog"
      },
      "delete": {
        "tags": ["Blogs"],
        "summary": "Xóa bài viết blog (Admin, soft delete)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Đã xóa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "operationId": "deleteBlog"
      },
      "get": {
        "tags": ["Blogs"],
        "summary": "Lấy bài viết blog theo ID (Public)",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết bài viết",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Blog"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "operationId": "getBlogById"
      }
    },
...
    "/api/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Xem chi tiết một liên hệ (Admin/Staff)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin chi tiết liên hệ",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Contact" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "operationId": "getContactById"
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Xóa liên hệ (Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Đã xóa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        },
        "operationId": "deleteContact"
      }
    },
...
    "/api/contacts/{id}/reply": {
      "patch": {
        "tags": ["Contacts"],
        "summary": "Phản hồi liên hệ (Admin/Staff)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["reply"],
                "properties": {
                  "reply": {
                    "type": "string",
                    "description": "Nội dung phản hồi"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã phản hồi",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Contact"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "operationId": "replyContact"
      }
    },
...
  }
}